+++ tools/tests/merger/OpenApiMerger.test.ts
@@ -9,10 +9,10 @@
 
 import OpenApiMerger from 'merger/OpenApiMerger'
 import fs from 'fs'
-import { Logger } from 'Logger'
+import { Logger, LogLevel } from 'Logger'
 
 test('merge()', () => {
-  const merger = new OpenApiMerger('./tools/tests/merger/fixtures/spec/', new Logger())
+  const merger = new OpenApiMerger('./tools/tests/merger/fixtures/spec/', new Logger(LogLevel.error))
   merger.merge('./tools/tests/merger/opensearch-openapi.yaml')
   expect(fs.readFileSync('./tools/tests/merger/fixtures/expected.yaml', 'utf8'))
     .toEqual(fs.readFileSync('./tools/tests/merger/opensearch-openapi.yaml', 'utf8'))